# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Quick Menu API
  version: "1.0.0"
  description: |
    Quick Menu backend API — QR-based menu, real-time orders (WebSocket), admin dashboards.
servers:
  - url: http://localhost:8080
    description: Local dev server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Restaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        timezone:
          type: string
        currency:
          type: string
        address:
          type: string
    Table:
      type: object
      properties:
        id:
          type: string
        restaurantId:
          type: string
        name:
          type: string
        qrUrl:
          type: string
          format: uri
...
    Dish:
      type: object
      properties:
        id:
          type: string
        restaurantId:
          type: string
        categoryId:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        imageUrl:
          type: string
        isAvailable:
          type: boolean
        prepTimeMins:
          type: integer
...
  parameters:
    restaurantId:
      name: restaurantId
      in: path
      required: true
      schema:
        type: string
...
    dishId:
      name: dishId
      in: path
      required: true
      schema:
        type: string
...
security:
  - bearerAuth: []
paths:
...
  /api/restaurants:
    post:
      summary: Create a restaurant (admin)
      tags: [Restaurants]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Restaurant'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
    get:
      summary: List restaurants (paginated)
      tags: [Restaurants]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: sort
          in: query
          schema:
            type: string
            default: id,asc
            description: "sort=property,asc|desc (repeatable)"
      responses:
        '200':
          description: Page of restaurants
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
                  pageable:
                    type: object
                  totalElements:
                    type: integer
...
  /api/restaurants/{restaurantId}/tables:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
    post:
      summary: Create a table (generates QR url)
      tags: [Tables]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
    get:
      summary: List tables (paginated)
      tags: [Tables]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: sort
          in: query
          schema:
            type: string
            default: id,asc
      responses:
        '200':
          description: Page of tables
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Table'
                  totalElements:
                    type: integer
...
  /api/{restaurantId}/dishes:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
    post:
      summary: Create dish (admin)
      tags: [Menu]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dish'
      responses:
        '201':
          description: Created dish
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
    get:
      summary: List dishes (paginated)
      tags: [Menu]
      parameters:
        - name: includeUnavailable
          in: query
          schema: { type: boolean, default: false }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: sort
          in: query
          schema:
            type: string
            default: name,asc
      responses:
        '200':
          description: Page of dishes
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dish'
                  totalElements:
                    type: integer
  /api/{restaurantId}/dishes/{dishId}:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
      - $ref: '#/components/parameters/dishId'
    patch:
      summary: Update dish (partial)
      tags: [Menu]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dish'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '404':
          description: Not found
  /api/{restaurantId}/dishes/{dishId}/availability:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
      - $ref: '#/components/parameters/dishId'
    patch:
      summary: Toggle availability (admin/staff)
      tags: [Menu]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isAvailable:
                  type: boolean
      responses:
        '200':
          description: Updated availability
          content:
            application/json:
              schema:
                type: object
                properties:
                  dishId:
                    type: string
                  isAvailable:
                    type: boolean
...